move DOM related content from index.js to getItems.js
